[@types/ember] move -private/handlebars into @ember/template
