+++ types/ember-data/tsconfig.json
@@ -36,6 +36,8 @@
             "@ember/service": ["ember__service"],
             "@ember/string": ["ember__string"],
             "@ember/string/*": ["ember__string/*"],
+            "@ember/template": ["ember__template"],
+            "@ember/template/*": ["ember__template/*"],
             "@ember/test": ["ember__test"],
             "@ember/test/*": ["ember__test/*"],
             "@ember/utils": ["ember__utils"],
