+++ types/ember__string/tsconfig.json
@@ -15,6 +15,8 @@
             "../"
         ],
         "paths": {
+            "@ember/template": ["ember__template"],
+            "@ember/template/*": ["ember__template/*"],
             "@ember/string": ["ember__string"],
             "@ember/string/*": ["ember__string/*"]
         },
@@ -23,7 +25,6 @@
         "forceConsistentCasingInFileNames": true
     },
     "files": [
-        "-private/handlebars.d.ts",
         "index.d.ts",
         "ember__string-tests.ts"
     ]
