+++ types/ember__template/-private/handlebars.d.ts
