+++ types/ember__test-helpers/tsconfig.json
@@ -39,6 +39,8 @@
             "@ember/service": ["ember__service"],
             "@ember/string": ["ember__string"],
             "@ember/string/*": ["ember__string/*"],
+            "@ember/template": ["ember__template"],
+            "@ember/template/*": ["ember__template/*"],
             "@ember/test": ["ember__test"],
             "@ember/test/*": ["ember__test/*"],
             "@ember/utils": ["ember__utils"],
